pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mdv"
version = "0.1.0"
description = "Markdown viewer"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4==4.13.4",
    "Jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.3",
    "mdit-py-plugins==0.4.2",
    "mdurl==0.1.2",
    "Pygments==2.19.1",
    "soupsieve==2.7",
    "typing_extensions==4.14.0",
    "Werkzeug==3.1.3",
]

[project.scripts]
mdv = "server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["templates/*", "static/*"]

